Update config files and add new scripts
This commit includes changes in configuration files to include "bluez" package in the Bluetooth configuration, and deactivating the "rofi" program. It also adds new options for "awesome-plasma" in the desktop manager session. Additionally, the commit introduces new scripts for managing AwesomeWM with KDE Plasma, and 'nirrofi.sh' to run selected packages from a list with rofi.